adiçao do controle de quantidade de execuçoes do script de execuçao da soluçao

+++ Script Execuçao/script_trafego_internet.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-#IMPLEMENTAR REGRAS CASO ALGUM SCRIPT FALHAR E NAO PERMITIR EXECUTAR SE O PROCESSO DA CHAMADA ANTERIOR AINDA ESTIVER EXECUTANDO!!!!
+qtd_processos=`ps aux | grep -c "script_trafego_internet.sh"`
 
-#Limpa arquivos processados
-rm /indicadores_internet/trafego_internet/access_processado.log 2>/tmp/erro
+if [ "$qtd_processos" -lt 4 ]
+then
 
-#Envia arquivo access.log pro Hadoop
-/usr/local/hadoop/bin/hadoop fs -put /var/log/squid3/access.log /indicadores_internet/trafego_internet/access_processado.log
+	#Limpa arquivos processados
+	rm /indicadores_internet/trafego_internet/access_processado.log 2>/tmp/erro
 
-#Executa o .jar no Hadoop e guarda os resultados
-/usr/local/hadoop/bin/hadoop jar /indicadores_internet/trafego_internet/TrafegoInternet.jar /indicadores_internet/trafego_internet/access_processado.log /indicadores_internet/trafego_internet
+	#Envia arquivo access.log pro Hadoop
+	/usr/local/hadoop/bin/hadoop fs -put /var/log/squid3/access.log /indicadores_internet/trafego_internet/access_processado.log
 
-#Executa script python pra cadastrar o setor aos usuarios no banco Postgres
-python /indicadores_internet/trafego_internet/cadastrarSetorUsuarios.py
+	#Executa o .jar no Hadoop e guarda os resultados
+	/usr/local/hadoop/bin/hadoop jar /indicadores_internet/trafego_internet/TrafegoInternet.jar /indicadores_internet/trafego_internet/access_processado.log /indicadores_internet/trafego_internet
 
+	#Executa script python pra cadastrar o setor aos usuarios no banco Postgres
+	python /indicadores_internet/trafego_internet/cadastrarSetorUsuarios.py
+
+fi
